requirements.txt
python-dotenv
Pillow==11.2.1
markdown>=3.4.4
pygments>=2.15.1
django-markdownx>=4.0.2
